[modularize] Separate object store client from `Worker`.
